refactoring + review code
