+++ package.json
@@ -22,9 +22,11 @@
     "@angular/platform-browser-dynamic": "~8.2.13",
     "@angular/router": "~8.2.13",
     "@kolkov/angular-editor": "^1.0.0-beta.4",
+    "@types/ramda": "^0.26.34",
     "bootstrap": "^4.3.1",
     "hammerjs": "^2.0.8",
     "ngx-bootstrap": "^5.2.0",
+    "ramda": "^0.26.1",
     "rxjs": "~6.4.0",
     "travis": "^0.1.1",
     "travis-cli": "^1.0.9",
