+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { UserInformationComponent } from './user-registry/components/user-information.component';
-import { UserExperianceResolverService } from './user-registry/user-experiance/service/user-experiance-resolver.service';
+import { UserExperianceResolverService } from './user-registry/user-project/service/user-experiance-resolver.service';
