+++ src/app/user-registry/components/user-information.component.html
@@ -1,3 +1,3 @@
 <div class="container">
-    <app-user-experiance></app-user-experiance>
+    <app-user-project></app-user-project>
 </div>
