--- src/app/user-registry/user-experiance/components/user-experiance.component.html
@@ -1,119 +1,0 @@
-<div class=" order-md-1">
-    <h4 class="mb-3">Project</h4>
-    <form [formGroup]="projectForm" (ngSubmit)="saveProject()">
-        <div class="mb-3">
-            <mat-form-field class="example-full-width">
-                <input type="text" placeholder="Intitulé du projet" matInput formControlName="entitle">
-
-            </mat-form-field>
-            <div *ngIf="entitle.invalid && (entitle.dirty || entitle.touched)" class="alert alert-danger">
-
-                <div *ngIf="entitle.errors.required">
-                    Intitulé du projet requis
-                </div>
-            </div>
-        </div>
-        <div class="mb-3">
-            <angular-editor [config]="config" formControlName="description"></angular-editor>
-            <div *ngIf="description.invalid && (description.dirty || description.touched)" class="alert alert-danger">
-
-                <div *ngIf="description.errors.required">
-                    Description du projet requis
-                </div>
-            </div>
-        </div>
-        <div class="mb-3">
-            <mat-form-field class="example-full-width">
-                <input type="text" placeholder="Nom de l'entreprise" matInput formControlName="entrepriseName">
-
-            </mat-form-field>
-            <div *ngIf="entrepriseName.invalid && (entrepriseName.dirty || entrepriseName.touched)"
-                class="alert alert-danger">
-
-                <div *ngIf="entrepriseName.errors.required">
-                    Nom de l'entreprise requis
-                </div>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-md-6 mb-3">
-                <mat-form-field class="example-full-width">
-                    <input matInput [min]="dateStartMin" [max]="dateStartMax" [matDatepicker]="pickerStart"
-                        placeholder="Du" formControlName="startDate">
-                    <mat-datepicker-toggle matSuffix [for]="pickerStart"></mat-datepicker-toggle>
-                    <mat-datepicker #pickerStart></mat-datepicker>
-                </mat-form-field>
-                <div *ngIf="startDate.invalid && (startDate.dirty || startDate.touched)" class="alert alert-danger">
-
-                    <div *ngIf="startDate.errors.required">
-                        Date début requis
-                    </div>
-                </div>
-            </div>
-            <div class="col-md-6 mb-3">
-                <mat-form-field class="example-full-width">
-                    <input matInput [min]="dateEndMin" [max]="dateEndMax" [matDatepicker]="pickerEnd" placeholder="Au"
-                        formControlName="endDate">
-                    <mat-datepicker-toggle matSuffix [for]="pickerEnd"></mat-datepicker-toggle>
-                    <mat-datepicker #pickerEnd></mat-datepicker>
-                </mat-form-field>
-                <div *ngIf="endDate.invalid && (endDate.dirty || endDate.touched)" class="alert alert-danger">
-
-                    <div *ngIf="endDate.errors.required">
-                        Date fin requis
-                    </div>
-                </div>
-            </div>
-        </div>
-
-
-
-        <div class="mb-3">
-            <mat-form-field class="example-full-width">
-                <mat-label>Role</mat-label>
-                <mat-select formControlName="role">
-                    <mat-option *ngFor="let role of roles" [value]="role">
-                        {{role}}
-                    </mat-option>
-                </mat-select>
-            </mat-form-field>
-            <div *ngIf="role.invalid && (role.dirty || role.touched)" class="alert alert-danger">
-
-                <div *ngIf="role.errors.required">
-                    Rôle requis
-                </div>
-            </div>
-        </div>
-
-
-        <div class="mb-3">
-            <mat-form-field class="example-full-width">
-                <mat-chip-list #chipList>
-                    <mat-chip *ngFor="let skill of skillsSelected" [selectable]="selectable" [removable]="removable"
-                        (removed)="remove(skill)">
-                        {{skill}}
-                        <mat-icon matChipRemove *ngIf="removable">cancel</mat-icon>
-                    </mat-chip>
-                    <input placeholder="Skills..." #skillInput [formControl]="skillCtrl" [matAutocomplete]="auto"
-                        [matChipInputFor]="chipList" [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
-                        [matChipInputAddOnBlur]="addOnBlur" (matChipInputTokenEnd)="add($event)">
-                </mat-chip-list>
-                <mat-autocomplete #auto="matAutocomplete" (optionSelected)="selected($event)">
-                    <mat-option *ngFor="let skill of filteredSkills | async" [value]="skill">
-                        {{skill}}
-                    </mat-option>
-                </mat-autocomplete>
-            </mat-form-field>
-            <div *ngIf="skills.invalid && (skillCtrl.dirty || skillCtrl.touched)" class="alert alert-danger">
-
-                <div *ngIf="skills.errors.required">
-                    Skill requis
-                </div>
-            </div>
-
-        </div>
-
-
-        <button class="btn btn-primary" type="submit">Enregistrer</button>
-    </form>
-</div>
