--- src/app/user-registry/user-experiance/components/user-experiance.component.scss
@@ -1,3 +1,0 @@
-.example-full-width {
-  width: 100%;
-}
