--- src/app/user-registry/user-experiance/model/entreprise.ts
@@ -1,4 +1,0 @@
-export interface Entreprise {
-    name: string;
-    address: string;
-}
