+++ src/app/user-registry/user-project/components/user-project.component.html
@@ -1,0 +1,111 @@
+<div class=" order-md-1">
+    <h4 class="mb-3">Projet</h4>
+    <form [formGroup]="userProjectForm" (ngSubmit)="saveUserProject()">
+        <div class="mb-3">
+            <mat-form-field class="w-100">
+                <input type="text" placeholder="Intitulé du projet" matInput formControlName="subject">
+
+            </mat-form-field>
+            <div *ngIf="userProjectForm.get('subject').invalid && (userProjectForm.get('subject').dirty || userProjectForm.get('subject').touched)" class="alert alert-danger">
+
+                <div *ngIf="userProjectForm.get('subject').errors.required">
+                    Intitulé du projet requis
+                </div>
+            </div>
+        </div>
+        <div class="mb-3">
+            <angular-editor [config]="config" formControlName="description"></angular-editor>
+            <div *ngIf="userProjectForm.get('description').invalid && (userProjectForm.get('description').dirty || userProjectForm.get('description').touched)" class="alert alert-danger">
+
+                <div *ngIf="userProjectForm.get('description').errors.required">
+                    Description du projet requis
+                </div>
+            </div>
+        </div>
+        <div class="mb-3">
+            <mat-form-field class="w-100">
+                <input type="text" placeholder="Nom de l'entreprise" matInput formControlName="enterpriseName">
+
+            </mat-form-field>
+            <div *ngIf="userProjectForm.get('enterpriseName').invalid && (userProjectForm.get('enterpriseName').dirty || userProjectForm.get('enterpriseName').touched)"
+                class="alert alert-danger">
+
+                <div *ngIf="userProjectForm.get('enterpriseName').errors.required">
+                    Nom de l'entreprise requis
+                </div>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col-md-6 mb-3">
+                <mat-form-field class="w-100">
+                    <input matInput [min]="minStartDate" [max]="maxStartDate" [matDatepicker]="pickerStart"
+                        placeholder="Du" formControlName="startDate">
+                    <mat-datepicker-toggle matSuffix [for]="pickerStart"></mat-datepicker-toggle>
+                    <mat-datepicker #pickerStart></mat-datepicker>
+                </mat-form-field>
+                <div *ngIf="userProjectForm.get('startDate').invalid && (userProjectForm.get('startDate').dirty || userProjectForm.get('startDate').touched)" class="alert alert-danger">
+
+                    <div *ngIf="userProjectForm.get('startDate').errors.required">
+                        Date début requis
+                    </div>
+                </div>
+            </div>
+            <div class="col-md-6 mb-3">
+                <mat-form-field class="w-100">
+                    <input matInput [min]="minEndDate" [max]="maxEndDate" [matDatepicker]="pickerEnd" placeholder="Au"
+                        formControlName="endDate">
+                    <mat-datepicker-toggle matSuffix [for]="pickerEnd"></mat-datepicker-toggle>
+                    <mat-datepicker #pickerEnd></mat-datepicker>
+                </mat-form-field>
+            </div>
+        </div>
+
+
+
+        <div class="mb-3">
+            <mat-form-field class="w-100">
+                <mat-label>Role</mat-label>
+                <mat-select formControlName="role">
+                    <mat-option *ngFor="let role of roles" [value]="role">
+                        {{role}}
+                    </mat-option>
+                </mat-select>
+            </mat-form-field>
+            <div *ngIf="userProjectForm.get('role').invalid && (userProjectForm.get('role').dirty || userProjectForm.get('role').touched)" class="alert alert-danger">
+
+                <div *ngIf="userProjectForm.get('role').errors.required">
+                    Rôle requis
+                </div>
+            </div>
+        </div>
+
+
+        <div class="mb-3">
+            <mat-form-field class="w-100">
+                <mat-chip-list #chipList>
+                    <mat-chip *ngFor="let skill of skillsSelected" [selectable]="true" [removable]="true"
+                        (removed)="removeSkill(skill)">
+                        {{skill}}
+                        <mat-icon matChipRemove >cancel</mat-icon>
+                    </mat-chip>
+                    <input placeholder="Skills..." #skillInput [formControl]="skillCtrl" [matAutocomplete]="autocompletionSkill"
+                        [matChipInputFor]="chipList" [matChipInputSeparatorKeyCodes]="separatorKeysCodes"
+                        [matChipInputAddOnBlur]="false" (matChipInputTokenEnd)="addSkill($event)">
+                </mat-chip-list>
+                <mat-autocomplete #autocompletionSkill="matAutocomplete" (optionSelected)="selectedSkillFromAoutocomplete($event)">
+                    <mat-option *ngFor="let skill of filteredSkills | async" [value]="skill">
+                        {{skill}}
+                    </mat-option>
+                </mat-autocomplete>
+            </mat-form-field>
+            <div *ngIf="userProjectForm.get('skills').invalid && (skillCtrl.dirty || skillCtrl.touched)" class="alert alert-danger">
+
+                <div *ngIf="userProjectForm.get('skills').errors.required">
+                    Skill requis
+                </div>
+            </div>
+        </div>
+
+        <button class="btn btn-primary" type="submit" [disabled]="userProjectForm.invalid">Enregistrer</button>
+    </form>
+</div>
