+++ src/app/user-registry/user-project/model/enterprise.ts
@@ -1,0 +1,5 @@
+export interface Enterprise {
+    id: number;
+    name: string;
+    address: string;
+}
